move metadata endpoints into their respective controllers
